+++ timezonefinder/helpers.py
@@ -2,6 +2,8 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 from math import asin, atan2, ceil, cos, degrees, radians, sin, sqrt
 
+from six.moves import range
+
 TIMEZONE_NAMES_FILE = 'timezone_names.json'